requirements.txt
streamlit
pandas
requests
mysql-connector-python
python-dotenv
